Hide delete post menu when the owner is the author of the post. #177

+++ lib/screen/parts/reaction/more.dart
@@ -2,6 +2,7 @@ import 'package:acceptable/acceptable.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bluesky/api/model/feed.dart';
+import 'package:flutter_bluesky/flutter_bluesky.dart';
 import 'package:flutter_bluesky/screen/parts/reaction.dart';
 import 'package:flutter_bluesky/util/post_util.dart';
 import 'package:flutter_bluesky/util/reaction_util.dart';
@@ -58,7 +59,7 @@ class MoreReaction extends AbstractReaction {
 
   @override
   Future<void> exec() async {
-    await popupMenu(context, [
+    List<PopupMenuItem> items = [
       // PopupMenuItem(
       //   child: Text(tr("menu.translate")),
       //   onTap: () => debugPrint("menu.translate"),
@@ -71,20 +72,23 @@ class MoreReaction extends AbstractReaction {
       //   child: Text(tr("menu.share")),
       //   onTap: () => debugPrint("menu.share"),
       // ),
-      // TODO #165
-      // PopupMenuItem(
-      //   child: Text(tr("menu.mute.thread")),
-      //   onTap: () => debugPrint("menu.mute.thread"),
-      // ),
-      PopupMenuItem(
-        child: Text(tr("report.post")),
-        onTap: () async => await PostUtil.report(context, post),
-      ),
-      // if the post user is the login user, delete post
-      PopupMenuItem(
+    ];
+    if (post.author.did == plugin.api.session.did) {
+      items.add(PopupMenuItem(
         child: Text(tr("delete.post")),
         onTap: () async => await PostUtil.delete(context, post),
-      ),
-    ]);
+      ));
+    } else {
+      // TODO #165
+      // items.add(PopupMenuItem(
+      //   child: Text(tr("mute.thread")),
+      //   onTap: () async => debugPrint("menu.mute.thread"),
+      // ));
+      items.add(PopupMenuItem(
+        child: Text(tr("report.post")),
+        onTap: () async => await PostUtil.report(context, post),
+      ));
+    }
+    await popupMenu(context, items);
   }
 }
